updated carousels with code duplication fix
